move bookTreeItemType to utils
