+++ extensions/notebook/src/common/utils.ts
@@ -12,7 +12,7 @@ import * as azdata from 'azdata';
 import * as crypto from 'crypto';
 import { notebookLanguages, notebookConfigKey, pinnedBooksConfigKey, AUTHTYPE, INTEGRATED_AUTH, KNOX_ENDPOINT_PORT, KNOX_ENDPOINT_SERVER } from './constants';
 import { IPrompter, IQuestion, QuestionTypes } from '../prompts/question';
-import { BookTreeItemFormat, BookTreeItemType } from '../book/bookTreeItem';
+import { BookTreeItemFormat } from '../book/bookTreeItem';
 import * as loc from './localizedConstants';
 
 const localize = nls.loadMessageBundle();
@@ -440,6 +440,20 @@ export function isBookItemPinned(notebookPath: string): boolean {
 		return true;
 	}
 	return false;
+}
+
+export enum BookTreeItemType {
+	Book = 'Book',
+	Notebook = 'Notebook',
+	Markdown = 'Markdown',
+	ExternalLink = 'ExternalLink',
+	providedBook = 'providedBook',
+	savedBook = 'savedBook',
+	unsavedNotebook = 'unsavedNotebook',
+	savedNotebook = 'savedNotebook',
+	pinnedNotebook = 'pinnedNotebook',
+	section = 'section',
+	savedBookNotebook = 'savedBookNotebook'
 }
 
 export function getNotebookType(book: BookTreeItemFormat): BookTreeItemType {
